tritonsrc: new kernel porting complete, rename to fwd_kernel
