+++ tritonsrc/flash.py
@@ -17,8 +17,7 @@ Extra Credits:
 import triton
 import triton.language as tl
 
-# from fwd_kernel import attn_fwd
-from new_fwd import attn_fwd
+from fwd_kernel import attn_fwd
 from bwd_preprocess import bwd_preprocess, bwd_preprocess_varlen
 from bwd_split_kernel import bwd_kernel_dk_dv, bwd_kernel_dq
 from dropout_rng import debug_fill_dropout_rng, debug_fill_dropout_rng_tensor
